+++ packages/fmc-autoform/inputs/MDSelect.jsx
@@ -30,83 +30,11 @@ AutoInput.MDSelect = React.createClass(
 
 	getInitialState()
 	{
-		let options = this.props.options ||
-			{},
-			open = false,
-			items = this.props.items || options.items || [],
-			selectedItem = this.props.value || this.props.selectedItem,
-			disabled = this.props.disabled || !items || !items.length,
-			errors = this.props.errors,
-			readOnly = this.props.readOnly || disabled;
-
-		if ( items != null && items.length )
-		{
-			let firstItem = items[ 0 ];
-			if( _.isString( firstItem ) )
-			{
-				items = items.sort();				
-			}
-			else if( firstItem.name != null )
-			{
-				items = items.sort( ( a, b ) => { return a.name > b.name ? 1 : -1 } )
-			}
-		}
-
-		return {
-			open,
-			items,
-			selectedItem,
-			disabled,
-			errors,
-			readOnly
-		};
+		return { open: false }
 	},
 
-	componentWillReceiveProps( props )
+	handleChange( newItem )
 	{
-		let options = props.options ||
-			{},
-			items = props.items || options.items || [],
-			selectedItem = props.value || props.selectedItem,
-			disabled = props.disabled || !items || !items.length,
-			errors = props.errors,
-			readOnly = props.readOnly || disabled;
-
-		if ( items != null && items.length )
-		{
-			let firstItem = items[ 0 ];
-			if( _.isString( firstItem ) )
-			{
-				items = items.sort();				
-			}
-			else if( firstItem.name != null )
-			{
-				items = items.sort( ( a, b ) => { return a.name > b.name ? 1 : -1 } )
-			}
-		}
-
-		this.setState(
-		{
-			items,
-			selectedItem,
-			disabled,
-			errors,
-			readOnly
-		} );
-		this.findSelectedItem();
-	},
-
-	componentWillMount()
-	{
-		this.findSelectedItem();
-	},
-
-	handleChange( newItem, event )
-	{
-		if ( event )
-		{
-			event.stopPropagation();
-		}
 		if ( this.props.options && this.props.options.onChange )
 		{
 			this.props.options.onChange( newItem );
@@ -133,180 +61,142 @@ AutoInput.MDSelect = React.createClass(
 
 	inputIsUsed( selectedItem )
 	{
-		return _.isObject( selectedItem ) ?
-			( selectedItem._id && selectedItem._id.length ) || ( selectedItem.name && selectedItem.name.length ) :
-			selectedItem && selectedItem.length;
+		if( _.isObject( selectedItem ) )
+		{
+			return (
+				( selectedItem._id != null && selectedItem._id.length > 0 ) ||
+				( selectedItem.name != null && selectedItem.name.length > 0 )
+			)
+		}
+		else if( _.isString( selectedItem ) )
+		{
+			return selectedItem != null && selectedItem.length > 0;
+		}
 	},
 
-	//search for the selected item in the provided list of items
-	// if it is found select it
-	// this if for the case where we have dependent fields
-	// for example location selector dependent on facility selector value
-	// we need to 
-	findSelectedItem()
+	setOpen( val )
 	{
-		let
-		{
-			items,
-			selectedItem
-		} = this.state;
-
-		if ( items != null && items.length > 0 && selectedItem != null )
-		{
-			let q = {},
-				checkField = null,
-				matchedItem = null;
-
-			if ( selectedItem._id != null )
-			{
-				checkField = '_id';
-			}
-			else if ( selectedItem.name != null )
-			{
-				checkField = 'name';
-				q = {
-					name: selectedItem.name
-				};
-			}
-
-			if ( checkField )
-			{
-				q[ checkField ] = selectedItem[ checkField ];
-				matchedItem = _.findWhere( items, q );
-				if ( matchedItem && !_.isEqual( matchedItem, selectedItem ) )
-				{
-					this.handleChange( matchedItem );
-				}
-			}
-		}
+		this.setState({ open:val });
 	},
 
 	render()
 	{
+		let { 
+			value, 
+			disabled, 
+			readOnly, 
+			errors, 
+			placeholder, 
+			description, 
+			items, 
+			view = PlainCard, //should perhaps be called Tile in the original schema context
+			clearOption 
+		} = this.props;
 
-		let options = this.props.options || {},
-			{
-				open,
-				items,
-				selectedItem,
-				disabled,
-				readOnly
-			} = this.state,
+		let ListTile = view;
 
-			errors = this.state.errors,
-			Card = this.props.itemView || options.view || PlainCard,
-			classes = this.props.classes || options.classes || '',
-			clearOption = this.props.clearOption || options.clearOption,
-			used = this.inputIsUsed( selectedItem );
+		if( items == null || items.length == 0 ) 
+		{
+			disabled = true;
+			readOnly = true;
+		}
 
-		//console.log( errors );
+		if( errors != null && errors.length > 0 )
+		{
+			invalid = true;
+		}
+
+		used = this.inputIsUsed( value );
+
+		let classes = ['input'];
+		if( invalid ) 
+		{
+			classes.push('invalid');
+		}
+		if( used )
+		{
+			classes.push('used');
+		}
+		if( this.state.open )
+		{
+			classes.push('open');
+		}
 
 		if ( readOnly )
 		{
 			return (
-				<div className = {"md-input md-select readonly dropdown "+classes+(disabled?" disabled":'')}>
-					<span className={"input"+(used?" used":'')+(errors?" invalid":'')}>
-	      				{used?<Card item={selectedItem}/>:<span>&nbsp;</span>}
-	      			</span>
+				<div className = {"md-input md-select readonly disabled dropdown"}>
+
+					<span className={ classes.join(' ') }>
+						{used?
+						<ListTile item = { value }/>
+						:<span>&nbsp;</span>}
+					</span>
+
 				    <span className="highlight"></span>
-      				<span className="bar"></span>
-      				<label>{this.props.placeholder}</label>
-		            {
-						errors?
-						<div className="helper-text">{ errors[0] }</div>
-						:null
-					}
+
+					<span className="bar"></span>
+
+					<label>{ placeholder }</label>
+
+			        {errors?
+					<div className="helper-text">{ errors[0] }</div>
+					:null}
+
 				</div>
 			)
 		}
 
-		return ( < div className = {
-				"md-input md-select dropdown" +
-				( this.state.open ? " open" : "" ) +
-				( classes ? ( " " + classes ) : "" )
-			}
-			tabIndex = "0"
-			onClick = {
-				() =>
-				{
-					if ( !this.state.open )
-					{
-						this.setState(
-						{
-							open: true
-						} )
-					}
-				}
-			}
-			onFocus = {
-				() =>
-				{
-					this.setState(
-					{
-						open: true
-					} )
-				}
-			}
-			onBlur = {
-				() =>
-				{
-					this.setState(
-					{
-						open: false
-					} )
-				}
-			} >
+		return (
+			<div className = {"md-input md-select dropdown" +( this.state.open ? " open" : "" )}
+				tabIndex = "0"
+				onClick = { () => { this.setOpen(true) } }
+				onFocus = { () => { this.setOpen(true) } }
+				onBlur = { () => { this.setOpen(false) } }>
 
-			< span className = {
-				"dropdown-toggle input" +
-				( used ? " used" : '' ) +
-				( this.state.open ? " focus" : '' ) +
-				( errors ? " invalid" : "" )
-			} >
-			{
-				used ? <Card item = { selectedItem }/> : <span>&nbsp;</span>
-			} < /span>
-			{
-				used ? <div className = "close-button" onClick={this.clearItem}>&times;</div> : null
-			}
-			<span className = "highlight"></span> < span className = "bar" > < /span> < label >
-			{
-				this.props.placeholder
-			} < /label>
+				<span className = { "dropdown-toggle "+classes.join(' ') }>
 
-			{
-				errors ?
-					<div className="helper-text">{ errors[0] }</div> : null
-			}
+				{used?
+				<ListTile item = { value }/>
+				:<span>&nbsp;</span>}
 
-			<ul className = "dropdown-menu">
-                    	{
-                    	this.props.description?
-                    	<li><div className="helper-text">{this.props.description}</div></li>
-                    	:null
-                    	}
+				</span>
 
-	                    {
-	                    	items.map( ( item, idx ) => {
-								return (
-		                        	<li 
-		                        		key = { idx+'-'+(item._id || item.name) } 
-		                        		className = "dropdown-menu-item" 
-		                        		onClick = { (event) => 
-		                        		{
-		                        			this.handleChange( item, event )
-		                        		}}>
+				{used?
+				<div className = "close-button" onClick={this.clearItem}>&times;</div>
+				:null}
 
-		                        		<Card item = { item } />
+				<span className = "highlight"></span>
+				<span className = "bar" > </span>
 
-		                        	</li>
-								)                    	
-	                    	})
-	                    }
+				<label>{ placeholder }</label>
 
-                    </ul>
+				{errors?
+				<div className="helper-text">{ errors[0] }</div>
+				:null}
 
-			< /div>
+				<ul className = "dropdown-menu">
+
+		        	{description?
+		        	<li><div className="helper-text">{ description }</div></li>
+		        	:null}
+
+		        	{items.map( ( item, idx ) => { 
+		        	/********************************************/
+		        	return (
+			    	<li key = { idx+'-'+(item._id || item.name) } 
+			    		className = "dropdown-menu-item" 
+			    		onClick = { () => { this.handleChange( item ) } }>
+
+			    		<ListTile item = { item } />
+
+			    	</li> )
+		        	/********************************************/
+			        })}
+
+	            </ul>
+
+			</div>
 		)
 	}
 } )